Moved ParametersFromProperties to parameters package and renamed custom CLI runner
